==> submit/KB부동산 가격지수/1.get_region_code.ipynb <==
from selenium import webdriver
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import time

urll = 'https://onland.kbstar.com/quics?page=C060737'

dir_for_shinhan = 'C:/Users/shic/Desktop/shinhan_simon/chromedriver/chromedriver.exe'
dir_for_seunghun = 'C:/Users/rsh15/Desktop/seunghuni/selenium/chromedriver.exe'

driver = webdriver.Chrome(dir_for_shinhan)
driver.maximize_window()
driver.refresh()

driver.get(urll)

big_region_code = []
big_region_name = []

select_big_region = driver.find_element_by_id('법정동대지역코드')
options = [x for x in select_big_region.find_elements_by_tag_name('option')]
for element in options:
    big_region_code.append(element.get_attribute('value'))
    big_region_name.append(element.get_attribute('text'))
    
big_region_data = pd.DataFrame({
    'big_region_code' : big_region_code,
    'big_region_name' : big_region_name
})
big_region_data

big_region_url = 'https://onland.kbstar.com/quics?page=C060737&%EB%B2%95%EC%A0%95%EB%8F%99%EB%8C%80%EC%A7%80%EC%97%AD%EC%BD%94%EB%93%9C={}'

small_region_list = []
for big_region in big_region_data.iterrows():
    
    small_region_code = []
    small_region_name = []

    driver.get(big_region_url.format(str(big_region[1]['big_region_code'])))
    
    time.sleep(3)
    
    select_small_region = driver.find_element_by_id('법정동중지역코드')
    options = [x for x in select_small_region.find_elements_by_tag_name('option')]
    for small_element in options:
        small_region_code.append(small_element.get_attribute('value'))
        small_region_name.append(small_element.get_attribute('text'))

    small_region_tmp = pd.DataFrame({
        'big_region_code'   : big_region[1]['big_region_code'],
        'big_region_name'   : big_region[1]['big_region_name'],
        'small_region_code' : small_region_code,
        'small_region_name' : small_region_name
    })
    small_region_list.append(small_region_tmp)

driver.close()

kb_regions = pd.concat(small_region_list,axis=0)
kb_regions['big_region_code'] = kb_regions['big_region_code'].astype(str)
kb_regions['small_region_code'] = kb_regions['small_region_code'].astype(str)
kb_regions.loc[kb_regions['small_region_name'] == '(전체)','small_region_code'] = ' '
kb_regions

dirr = 'C:/Users/shic/Desktop/crawler_project/data_save/kb_land/'
kb_regions.to_csv(dirr+'region_code.csv',index=False)
